fix can't remove files error

+++ main.py
@@ -1,5 +1,6 @@
 import os
 import configparser
+import glob
 
 config = configparser.ConfigParser()
 config.read('./configs/configs.ini')
@@ -35,6 +36,8 @@ if not os.path.isdir(logs_path):
         os.mkdir(f'{os.getcwd()}/logs')
 elif os.path.isdir(logs_path) and len(os.listdir(logs_path)) > 0:
         os.system(f'rm -r {os.getcwd()}/logs/*log*')
-        os.system(f'rm -r {os.getcwd()}/logs/*txt')
+        files = glob.glob("*.txt")
+        if files:
+                os.system(f'rm -r {os.getcwd()}/logs/*txt')
 # run batch file
 os.system(f"sbatch --job-name={jobName} --output={output} --time={time} --mem-per-cpu={memPerCpu} --nodes={nodes} --ntasks={ntasks} runBatch.sh")
